+++ elfo/tests/message_macro.rs
@@ -1,7 +1,7 @@
 use serde::Serialize;
 use static_assertions::*;
 
-use elfo::{message, Message, Request};
+use elfo::{message, set_protocol, Message, Request};
 
 #[message]
 struct SimpleMessage {}
@@ -11,6 +11,8 @@ assert_not_impl_all!(SimpleMessage: Request);
 #[message(ret = ())]
 struct SimpleRequest {}
 assert_impl_all!(SimpleRequest: Message, Request);
+assert_impl_all!(<SimpleRequest as Request>::Wrapper: Message);
+assert_not_impl_all!(<SimpleRequest as Request>::Wrapper: Request);
 
 #[message(part)]
 struct SimpleMessagePart(u32);
@@ -21,6 +23,38 @@ assert_not_impl_all!(SimpleMessagePart: Message, Request);
 struct TransparentMessagePart(u32);
 assert_impl_all!(TransparentMessagePart: std::fmt::Debug, Serialize);
 assert_not_impl_all!(TransparentMessagePart: Message, Request);
+
+#[message(protocol = "override")]
+struct SimpleMessageWithOverridedProtocol {}
+
+#[message(protocol = "override", ret = ())]
+struct SimpleRequestWithOverridedProtocol {}
+
+mod one {
+    use super::*;
+
+    set_protocol!("one");
+
+    #[message]
+    pub struct SimpleMessage {}
+
+    #[message(ret = ())]
+    pub struct SimpleRequest {}
+
+    pub mod two {
+        use super::*;
+
+        #[message]
+        pub struct SimpleMessage2 {}
+    }
+
+    pub mod three {
+        elfo::set_protocol!("three");
+
+        #[elfo::message]
+        pub struct SimpleMessage {}
+    }
+}
 
 #[test]
 fn transparent() {
@@ -35,6 +69,18 @@ fn transparent() {
 fn name() {
     assert_eq!(SimpleMessage {}.name(), "SimpleMessage");
     assert_eq!(SimpleRequest {}.name(), "SimpleRequest");
+    assert_eq!(
+        <SimpleRequest as Request>::Wrapper::from(()).name(),
+        "SimpleRequest::Response"
+    );
+    assert_eq!(
+        SimpleMessageWithOverridedProtocol {}.name(),
+        "SimpleMessageWithOverridedProtocol"
+    );
+    assert_eq!(one::SimpleMessage {}.name(), "SimpleMessage");
+    assert_eq!(one::SimpleRequest {}.name(), "SimpleRequest");
+    assert_eq!(one::two::SimpleMessage2 {}.name(), "SimpleMessage2");
+    assert_eq!(one::three::SimpleMessage {}.name(), "SimpleMessage");
     assert_eq!(elfo::messages::Ping::default().name(), "Ping");
 }
 
@@ -42,5 +88,23 @@ fn name() {
 fn protocol() {
     assert_eq!(SimpleMessage {}.protocol(), "elfo");
     assert_eq!(SimpleRequest {}.protocol(), "elfo");
+    assert_eq!(
+        <SimpleRequest as Request>::Wrapper::from(()).protocol(),
+        "elfo"
+    );
+    assert_eq!(SimpleMessageWithOverridedProtocol {}.protocol(), "override");
+    assert_eq!(SimpleRequestWithOverridedProtocol {}.protocol(), "override");
+    assert_eq!(
+        <SimpleRequestWithOverridedProtocol as Request>::Wrapper::from(()).protocol(),
+        "override"
+    );
+    assert_eq!(one::SimpleMessage {}.protocol(), "one");
+    assert_eq!(one::SimpleRequest {}.protocol(), "one");
+    assert_eq!(
+        <one::SimpleRequest as Request>::Wrapper::from(()).protocol(),
+        "one"
+    );
+    assert_eq!(one::two::SimpleMessage2 {}.protocol(), "one");
+    assert_eq!(one::three::SimpleMessage {}.protocol(), "three");
     assert_eq!(elfo::messages::Ping::default().protocol(), "elfo-core");
 }