+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - network: initial implementation of distribution.
 - macros: support `#[message(protocol = "X")]`.
+- macros: add `set_protocol!`.
 
 ### Changed
 - **BREAKING** errors: replace `RequestError::Closed(envelope)` with `Failed`.